agent/requirements.txt
firebase-admin==6.5.0
psutil==5.9.8
pyinstaller==6.5.0
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
send2trash==1.8.3
